refactor(template): fold Storefront MCP helpers into storefront.ts
Move the MCP client and catalog/product/policy functions out of the
non-standard storefront-mcp.ts into storefront.ts (reusing its store-domain
const and i18n imports), repoint tool imports, and trim comments to the
load-bearing ones.
