+++ apps/template/lib/shopify/storefront.ts
@@ -86,3 +86,161 @@ export const storefront = {
     return { data, errors };
   },
 };
+
+// Storefront MCP — the store's hosted agentic-commerce tools over JSON-RPC 2.0 (/api/mcp).
+const MCP_ENDPOINT = `https://${SHOPIFY_STORE_DOMAIN}/api/mcp`;
+const UCP_AGENT_PROFILE_URL = process.env.UCP_AGENT_PROFILE_URL;
+
+let mcpRpcId = 0;
+
+export async function callStorefrontMcp<T>(
+  tool: string,
+  args: Record<string, unknown>,
+): Promise<T> {
+  const meta = UCP_AGENT_PROFILE_URL
+    ? { "ucp-agent": { profile: UCP_AGENT_PROFILE_URL } }
+    : undefined;
+
+  const response = await fetch(MCP_ENDPOINT, {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      id: ++mcpRpcId,
+      jsonrpc: "2.0",
+      method: "tools/call",
+      params: { arguments: { ...(meta ? { meta } : {}), ...args }, name: tool },
+    }),
+  });
+
+  if (!response.ok) {
+    throw new Error(`Storefront MCP ${tool}: HTTP ${response.status}`);
+  }
+
+  const json = (await response.json()) as {
+    error?: { message?: string };
+    result?: {
+      content?: Array<{ text?: string; type?: string }>;
+      isError?: boolean;
+      structuredContent?: unknown;
+    };
+  };
+
+  if (json.error) {
+    throw new Error(`Storefront MCP ${tool}: ${json.error.message ?? "request failed"}`);
+  }
+
+  const result = json.result;
+  const text = result?.content?.find((c) => c.type === "text")?.text;
+
+  if (result?.isError) {
+    throw new Error(`Storefront MCP ${tool}: ${text ?? "tool returned an error"}`);
+  }
+
+  // Some servers return the payload as a JSON text block rather than structuredContent.
+  if (result?.structuredContent !== undefined) return result.structuredContent as T;
+  if (text) return JSON.parse(text) as T;
+
+  throw new Error(`Storefront MCP ${tool}: empty response`);
+}
+
+export interface McpMoney {
+  amount: number; // minor units (52800 = 528.00)
+  currency: string;
+}
+
+export interface McpCatalogProduct {
+  categories?: Array<{ name?: string }>;
+  description?: { html?: string };
+  id: string;
+  media?: Array<{ alt_text?: string; url?: string }>;
+  price_range?: { max?: McpMoney; min?: McpMoney };
+  tags?: string[];
+  title: string;
+  variants?: Array<{
+    availability?: { available?: boolean };
+    id: string;
+    media?: Array<{ url?: string }>;
+    price?: McpMoney;
+    title?: string;
+  }>;
+}
+
+export interface McpCatalogSearchResult {
+  instructions?: string;
+  pagination?: { cursor?: string; has_next_page?: boolean };
+  products?: McpCatalogProduct[];
+}
+
+export async function searchCatalog(params: {
+  intent?: string;
+  limit?: number;
+  locale?: string;
+  query: string;
+}): Promise<McpCatalogSearchResult> {
+  const { intent, limit = 10, locale = defaultLocale, query } = params;
+
+  const context: Record<string, unknown> = {
+    address_country: getCountryCode(locale),
+    language: getLanguageCode(locale),
+  };
+  if (intent) context.intent = intent;
+
+  return callStorefrontMcp<McpCatalogSearchResult>("search_catalog", {
+    catalog: { context, pagination: { limit }, query },
+  });
+}
+
+// get_product_details encodes money as major-unit strings with a sibling currency,
+// unlike search_catalog's minor-unit McpMoney objects.
+export interface McpProductDetails {
+  description?: string;
+  image_url?: string | null;
+  images?: Array<{ alt_text?: string; url?: string }>;
+  options?: Array<{ name?: string; values?: string[] }>;
+  price_range?: { currency?: string; max?: string; min?: string };
+  product_id: string;
+  selectedOrFirstAvailableVariant?: {
+    available?: boolean;
+    currency?: string;
+    price?: string;
+    title?: string;
+    variant_id?: string;
+  };
+  title: string;
+  total_variants?: number;
+}
+
+export async function getCatalogProduct(params: {
+  locale?: string;
+  options?: Record<string, string>;
+  productId: string;
+}): Promise<McpProductDetails | undefined> {
+  const { locale = defaultLocale, options, productId } = params;
+
+  const result = await callStorefrontMcp<{ product?: McpProductDetails }>("get_product_details", {
+    country: getCountryCode(locale),
+    language: getLanguageCode(locale),
+    product_id: productId,
+    ...(options ? { options } : {}),
+  });
+  return result.product;
+}
+
+export interface McpPolicyAnswer {
+  answer?: string;
+  question?: string;
+  sources?: Array<{ title?: string; url?: string }>;
+}
+
+export async function searchShopPoliciesAndFaqs(params: {
+  context?: string;
+  query: string;
+}): Promise<McpPolicyAnswer[]> {
+  const { context, query } = params;
+
+  const result = await callStorefrontMcp<McpPolicyAnswer | McpPolicyAnswer[]>(
+    "search_shop_policies_and_faqs",
+    { query, ...(context ? { context } : {}) },
+  );
+  return Array.isArray(result) ? result : [result];
+}
